renaming analysis file
